PP-5837 clean up connector code
- clean up connector code of any transaction related searches / events /
  single transaction views
- remove/update tests accordingly
