+++ test/cypress/test.env
@@ -20,4 +20,3 @@ STRIPE_HOST=localhost
 STRIPE_PORT=8000
 GOCARDLESS_TEST_OAUTH_BASE_URL=http://localhost:8000
 GOCARDLESS_TEST_CLIENT_ID=testClientId
-FEATURE_USE_LEDGER_PAYMENTS=true
