+++ test/test.env
@@ -23,4 +23,3 @@ SELFSERVICE_URL=https://selfservice.pymnt.localdomain
 ZENDESK_URL=https://govuk.zendesk.com/api/v2
 STRIPE_HOST=localhost
 STRIPE_PORT=8000
-FEATURE_USE_LEDGER_PAYMENTS=true
